+++ bulbs/content/tasks.py
@@ -90,10 +90,12 @@ def post_article(content, body, fb_page_id, fb_api_url, fb_token_path, fb_dev_mo
             Error in posting Instant Article.\n
             Content ID: {0}\n
             IA ID: {1}\n
-            Status Code: {2}'''.format(
+            Status Code: {2}
+            Request: {3}'''.format(
                 content.id,
                 content.instant_article_id,
-                post.status_code))
+                post.status_code,
+                post.__dict__))
         return
 
     # Poll for status of article
@@ -113,10 +115,12 @@ def post_article(content, body, fb_page_id, fb_api_url, fb_token_path, fb_dev_mo
                 Error in getting status of Instant Article.\n
                 Content ID: {0}\n
                 IA ID: {1}\n
-                Status Code: {2}'''.format(
+                Status Code: {2}
+                Request: {3}'''.format(
                     content.id,
                     content.instant_article_id,
-                    status.status_code))
+                    status.status_code,
+                    status.__dict__))
             return
 
         response = status.json().get('status')
@@ -143,10 +147,12 @@ def delete_article(content, fb_api_url, fb_token_path):
             Error in deleting Instant Article.\n
             Content ID: {0}\n
             IA ID: {1}\n
-            Status Code: {2}'''.format(
+            Status Code: {2}
+            Request: {3}'''.format(
                 content.id,
                 content.instant_article_id,
-                delete.status_code))
+                delete.status_code,
+                delete.__dict__))
     else:
         status = delete.json().get('success')
         if bool(status) is not True: