+++ bulbs/videos/migrations/0002_auto_20160613_1116.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('videos', '0001_initial'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='videohubvideo',
+            name='id',
+            field=models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True),
+        ),
+    ]